mark old run.py script for deprecation and rename run2.py to run.py
