+++ Makefile
@@ -69,7 +69,7 @@ TUMOR_FILE:=tumor_pileups.txt
 NORMAL_FILE:=normal_pileups.txt
 MARKERS:=data/markers/GRCh37.autosomes.phase3_shapeit2_mvncall_integrated.20130502.SNV.genotype.sselect_v4_MAF_0.4_LD_0.8.txt
 run:
-	python run2.py \
+	python run.py \
 	--threads "$(THREADS)" \
 	--tumors "$(NUM_TUMORS)" \
 	--normals "$(NUM_NORMALS)" \
