+++ run.py
@@ -1,38 +1,38 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 """
-Run the Conpair scripts in parallel for all combinations of tumor and normal pileups
+Import the Conpair functionality from its scripts and modules and run it in parallel for all samples in the lists
 """
-import sys
+import csv
 import os
-import subprocess as sp
+import sys
+import datetime
 import itertools
 from multiprocessing import Pool
-import datetime
-import re
 import json
+import argparse
 
+# import the script as a module from the other directory
 THIS_DIR = os.path.dirname(os.path.realpath(__file__))
 SCRIPT_DIR = os.path.join(THIS_DIR, "scripts")
-verify_concordance_script = os.path.join(SCRIPT_DIR, 'verify_concordance.py')
-estimate_tumor_normal_contamination_script = os.path.join(SCRIPT_DIR, 'estimate_tumor_normal_contamination.py')
+sys.path.insert(0, SCRIPT_DIR)
+import verify_concordance2
+sys.path.pop(0)
 
-def run_command(args):
-    """
-    Helper function to run a shell command easier
-    Parameters
-    ----------
-    args: list
-        a list of shell args to execute
-    """
-    process = sp.Popen(args, stdout = sp.PIPE, stderr = sp.PIPE, universal_newlines = True)
-    proc_stdout, proc_stderr = process.communicate()
-    returncode = process.returncode
-    proc_stdout = proc_stdout.strip()
-    proc_stderr = proc_stderr.strip()
-    return(returncode, proc_stdout, proc_stderr)
+# need to find a default set of targets to use
+default_marker_file = os.path.join(THIS_DIR, 'data', 'markers', 'GRCh37.autosomes.phase3_shapeit2_mvncall_integrated.20130502.SNV.genotype.sselect_v4_MAF_0.4_LD_0.8.txt')
 
-def run_conpair(tumor_pileup, normal_pileup, actions_list):
+# default parameters to run concordance with
+concordance_params = {
+'min_mapping_quality': 10,
+'normal_homozygous_markers_only': False,
+'markers': default_marker_file,
+'outfile': '-',
+'min_cov': 10,
+'min_base_quality': 20
+}
+
+def run_conpair(tumor_pileup, normal_pileup, actions_list, concordance_params):
     """
     Run the Conpair wrapper scripts on the tumor normal pair and collect the results
     """
@@ -49,100 +49,35 @@ def run_conpair(tumor_pileup, normal_pileup, actions_list):
             'normal_pileup': normal_pileup
         }
     }
-
     if 'concordance' in actions_list:
-        command = [ 'python2.7',  verify_concordance_script, '--tumor_pileup', tumor_pileup, '--normal_pileup', normal_pileup ]
-        returncode, proc_stdout, proc_stderr = run_command(command)
-
+        try:
+            concordance, num_markers_used, num_total_markers = verify_concordance2.main(
+            tumor_pileup = tumor_pileup,
+            normal_pileup = normal_pileup,
+            min_mapping_quality = concordance_params['min_mapping_quality'],
+            normal_homozygous_markers_only = concordance_params['normal_homozygous_markers_only'],
+            markers = concordance_params['markers'],
+            min_cov = concordance_params['min_cov'],
+            min_base_quality = concordance_params['min_base_quality'],
+            outfile = concordance_params['outfile']
+            )
+        except ZeroDivisionError:
+            print('WARNING: There are no shared markers between the tumor and the normal samples that meet the specified coverage requirements; tumor_pileup: {}, normal_pileup: {}'.format(tumor_pileup, normal_pileup))
+            concordance = None
+            num_markers_used = None
+            num_total_markers = None
         result['concordance'] = {}
-        result['concordance']['returncode'] = returncode
-        result['concordance']['stdout'] = proc_stdout
-        result['concordance']['stderr'] = proc_stderr
-        result['concordance']['time'] = str((datetime.datetime.now() - timestart).seconds)
-
-        concordance = None
-        if len(proc_stdout.split()) > 0:
-            try:
-                concordance = float(proc_stdout.split()[0].strip())
-            except:
-                pass
+        result['concordance']['time'] = (datetime.datetime.now() - timestart).seconds
         result['concordance']['concordance'] = concordance
+        result['concordance']['num_markers_used'] = num_markers_used
+        result['concordance']['num_total_markers'] = num_total_markers
 
-    if 'contamination' in actions_list:
-        command = [ 'python2.7',  estimate_tumor_normal_contamination_script, '--tumor_pileup', tumor_pileup, '--normal_pileup', normal_pileup ]
-        returncode, proc_stdout, proc_stderr = run_command(command)
-
-        result['contamination'] = {}
-        result['contamination']['returncode'] = returncode
-        result['contamination']['stdout'] = proc_stdout
-        result['contamination']['stderr'] = proc_stderr
-        result['contamination']['time'] = str((datetime.datetime.now() - timestart).seconds)
-
-        normal_contamination = None
-        tumor_contamination = None
-        matches = re.findall(r'[0-9]+.[0-9]+', proc_stdout)
-        if len(matches) > 0:
-            try:
-                normal_contamination = float(matches[0])
-                tumor_contamination = float(matches[1])
-            except:
-                pass
-
-        result['contamination']['normal_contamination'] = normal_contamination
-        result['contamination']['tumor_contamination'] = tumor_contamination
-
-    result['time'] = str((datetime.datetime.now() - timestart).seconds)
-
+    result['time'] = (datetime.datetime.now() - timestart).seconds
     return(result)
 
-def main():
+def print_to_console(timestop, num_pairs, time_taken):
     """
-    run.py <num_threads> <num_tumors> <num_normals> concordance,contamination
     """
-    # check if cli args were passed
-    args = sys.argv[1:]
-    if len(args) > 0:
-        num_threads = int(args[0])
-        num_tumors = int(args[1])
-        num_normals = int(args[2])
-        actions = args[3] # concordance,contamination
-    else:
-        num_threads = 8
-        num_tumors = 1
-        num_normals = 1
-        actions = "concordance,contamination"
-
-    actions_list = actions.split(',')
-
-    timestart = datetime.datetime.now()
-
-    # load the paths to pileups
-    tumor_pileup_file = "tumor_pileups.txt"
-    normal_pileup_file = "normal_pileups.txt"
-    with open(tumor_pileup_file) as fin:
-        all_tumor_pileups  = [ line.strip() for line in fin if line.strip() != '' ]
-    with open(normal_pileup_file) as fin:
-        all_normal_pileups  = [ line.strip() for line in fin if line.strip() != '' ]
-
-    tumor_pileups = all_tumor_pileups[0:num_tumors]
-    normal_pileups = all_normal_pileups[0:num_normals]
-
-    # get all combinations of both lists of pileups
-    pairs = list(itertools.product(tumor_pileups, normal_pileups))
-
-    # get the first one for testing
-    # run the Copair scripts in parallele in a multi-thread process pool
-    pool = Pool(num_threads)
-    results = [ pool.apply_async(run_conpair, args=(tumor_pileup, normal_pileup, actions_list)) for tumor_pileup, normal_pileup in pairs ] # [0:num_pairs]
-    output = [ p.get() for p in results ]
-
-    timestop = datetime.datetime.now()
-    time_taken = str((timestop - timestart).seconds)
-    num_pairs = len(output)
-    num_tumors = len(tumor_pileups)
-    num_normals = len(normal_pileups)
-    actions_str = '.'.join(actions_list)
-
     # print to console
     print("[{timestamp}] {num_pairs} pairs processed in {time_taken}s".format(
     timestamp = timestop,
@@ -150,15 +85,18 @@ def main():
     time_taken = time_taken
     ))
 
-    # save the output log
-    log_file = "output.{}pairs.{}t.{}n.{}thread.{}s.{}.json".format(
-        num_pairs,
-        num_tumors,
-        num_normals,
-        num_threads,
-        time_taken,
-        timestart.strftime('%Y-%m-%d_%H-%M-%S')
-    )
+def save_JSON_output(output, timestart, time_taken, num_pairs, num_tumors, num_normals, num_threads, actions_list, log_file = None):
+    """
+    """
+    if not log_file:
+        log_file = "output.{}pairs.{}t.{}n.{}thread.{}s.{}.json".format(
+            num_pairs,
+            num_tumors,
+            num_normals,
+            num_threads,
+            time_taken,
+            timestart.strftime('%Y-%m-%d_%H-%M-%S')
+        )
     log = {
     'output': output,
     'time': time_taken,
@@ -171,13 +109,109 @@ def main():
     with open(log_file, "w") as fout:
         json.dump(log, fout, indent = 4)
 
-    # save benchmarks
-    with open("benchmarks.tsv", "a") as fout:
+def save_benchmarks(num_threads, time_taken, num_pairs, num_tumors, num_normals, actions_str, write_mode = "a", output_file = "benchmarks.tsv"):
+    """
+    """
+    with open(output_file, write_mode) as fout:
         line = '\t'.join([str(num_threads), str(time_taken), str(num_pairs), str(num_tumors), str(num_normals), actions_str]) + '\n'
         fout.write(line)
 
+def save_table_output(output):
+    """
+    """
+    has_concordance = any([ 'concordance' in item for item in output ])
+    has_contamination = any([ 'contamination' in item for item in output ])
 
+    if has_concordance:
+        conc_out = open('concordance.tsv', "w")
+        conc_fieldnames = ['concordance', 'num_markers_used', 'num_total_markers', 'tumor', 'normal', 'tumor_pileup', 'normal_pileup']
+        conc_writer = csv.DictWriter(conc_out, delimiter = '\t', fieldnames = conc_fieldnames)
+        conc_writer.writeheader()
+    if has_contamination:
+        cont_out = open('concordance.tsv', "w")
+        cont_fieldnames = ['tumor', 'normal', 'tumor_pileup', 'normal_pileup']
+        cont_writer = csv.DictWriter(cont_out, delimiter = '\t', fieldnames = cont_fieldnames)
+        cont_writer.writeheader()
 
+    for item in output:
+        if 'concordance' in item:
+            row = {
+            'tumor': item['pair']['tumor'],
+            'tumor_pileup': item['pair']['tumor_pileup'],
+            'normal': item['pair']['normal'],
+            'normal_pileup': item['pair']['normal_pileup'],
+            'concordance': item['concordance']['concordance'],
+            'num_total_markers': item['concordance']['num_total_markers'],
+            'num_markers_used': item['concordance']['num_markers_used']
+            }
+            conc_writer.writerow(row)
+
+    if has_concordance:
+        conc_out.close()
+    if has_contamination:
+        cont_out.close()
+
+def main(**kwargs):
+    """
+    Main control function for the script
+    """
+    num_threads = kwargs.pop('num_threads', 4)
+    num_tumors = kwargs.pop('num_tumors', 1)
+    num_normals = kwargs.pop('num_normals', 1)
+    actions = kwargs.pop('actions', "concordance,contamination")
+    tumor_file = kwargs.pop('tumor_file', "tumors.txt")
+    normal_file = kwargs.pop('normal_file', "normals.txt")
+    markers = kwargs.pop('markers', default_marker_file)
+
+    concordance_params['markers'] = markers
+
+    actions_list = actions.split(',')
+
+    timestart = datetime.datetime.now()
+
+    # load the paths to pileups
+    with open(tumor_file) as fin:
+        all_tumor_pileups  = [ line.strip() for line in fin if line.strip() != '' ]
+    with open(normal_file) as fin:
+        all_normal_pileups  = [ line.strip() for line in fin if line.strip() != '' ]
+
+    tumor_pileups = all_tumor_pileups[0:num_tumors]
+    normal_pileups = all_normal_pileups[0:num_normals]
+
+    # get all combinations of both lists of pileups
+    pairs = list(itertools.product(tumor_pileups, normal_pileups))
+
+    # run the Copair scripts in parallele in a multi-thread process pool
+    pool = Pool(num_threads)
+    results = [ pool.apply_async(run_conpair, args=(tumor_pileup, normal_pileup, actions_list, concordance_params)) for tumor_pileup, normal_pileup in pairs ]
+    output = [ p.get() for p in results ]
+
+    timestop = datetime.datetime.now()
+    time_taken = (timestop - timestart).seconds
+    num_pairs = len(output)
+    num_tumors = len(tumor_pileups)
+    num_normals = len(normal_pileups)
+
+    print_to_console(timestop, num_pairs, time_taken)
+    save_JSON_output(output, timestart, time_taken, num_pairs, num_tumors, num_normals, num_threads, actions_list)
+    save_benchmarks(num_threads, time_taken, num_pairs, num_tumors, num_normals, '.'.join(actions_list))
+    save_table_output(output)
+
+def parse():
+    """
+    Parse the command line options
+    """
+    parser = argparse.ArgumentParser(description = 'Generate cBio Portal metadata files from various input files')
+    parser.add_argument('-t', '--threads', dest = 'num_threads', default = 4, type = int, help = 'The number of CPU threads to use')
+    parser.add_argument('--tumors', dest = 'num_tumors', default = 1, type = int, help = 'The number of tumor samples to use from the list')
+    parser.add_argument('--normals', dest = 'num_normals', default = 1, type = int,  help = 'The number of normal samples to use from the list')
+    parser.add_argument('--actions', dest = 'actions', default = 'concordance,contamination', help = 'A comma separate list of actions to run')
+    parser.add_argument('--tumor-file', dest = 'tumor_file', default = "tumors.txt", help = 'File with a list filepaths to the pileups of the tumor samples to use')
+    parser.add_argument('--normal-file', dest = 'normal_file', default = "normals.txt", help = 'File with a list filepaths to the pileups of the normal samples to use')
+    parser.add_argument('--markers', dest = 'markers', default = default_marker_file, help = 'Markers to use for analysis')
+
+    args = parser.parse_args()
+    main(**vars(args))
 
 if __name__ == '__main__':
-    main()
+    parse()
